refactor(app): refactoring project
Refactoring model, repository and test for TextAnalyzer.

+++ src/main/java/dev/ajdinkomic/textanalyzer/model/TextAnalyzer.java
@@ -8,6 +8,6 @@ public record TextAnalyzer(
         @NotNull
         UserInput userInput,
         HashMap<Character, Integer> analysisResult,
-        Long analysisDuration
+        Double analysisDuration
 ) {
 }
